Adding MOE configurations for DFP Showcase application. 	Change on 2012/03/09 by eleichtenschl <eleichtenschl@google.com> ------------- This automated CL converts BUILD files to the official formatting, by running them through Buildifier. For background, see [] and [].
This CL was generated with buildifier's "list sorting" mode disabled.
If deps or other lists need sorting, that will happen in a followup batch.

TESTING:

This CL was tested by comparing the output of '[] query' before
and after editing dir/BUILD. The changes in this CL leave that output
unchanged except for line number information, meaning that the
BUILD file has *exactly* the same meaning to [] that it did before
reformatting.

If any files in this CL are edited by other CLs before this can be
submitted, so that a sync is needed, Rosie will remove those files
from this CL (revert) rather than attempt a merge.
We'll pick up those files next time.
	Change on 2012/10/17 by rsc <rsc@google.com>
-------------
Adding tabbed view to start separating basic and advanced features in DFP Showcase.
	Change on 2013/01/30 by eleichtenschl <eleichtenschl@google.com>
-------------
Updating DFP Showcase v2 to use Bundled and Premium categorizations.
	Change on 2013/03/13 by eleichtenschl <eleichtenschl@google.com>
-------------
Removing the feature to mark a list item as premium. We are now going with separate tabs to denote whether or not a feature is premium.
	Change on 2013/03/15 by eleichtenschl <eleichtenschl@google.com>
-------------
Refactoring DFP Showcase to use an enum instead of an array of constants.
	Change on 2013/03/21 by eleichtenschl <eleichtenschl@google.com>
-------------
Adds Toast messages to DfpShowcase for AdMob callbacks.
	Change on 2013/03/26 by eleichtenschl <eleichtenschl@google.com>
-------------
Fixing AdFormat definition bug.
	Change on 2013/03/27 by eleichtenschl <eleichtenschl@google.com>
-------------
Updating DFP Showcase to:

- support 320dp-wide devices for the 360x50 sized samples.
- change size of image carousel and image with google +1 to 300x250
- change image animation to use interstitial
- pass "screenWidth" and "screenHeight" extras parameters to DfpExtras for interstitials. It is used by the creative to serve the right asset.
	Change on 2013/03/28 by eleichtenschl <eleichtenschl@google.com>
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=44559034